Extract modals from products views
