+++ saleor/static/dashboard-next/products/components/ProductVariantPage/ProductVariantPage.tsx
@@ -11,7 +11,6 @@ import Toggle from "../../../components/Toggle";
 import { UserError } from "../../../types";
 import { ProductVariant } from "../../types/ProductVariant";
 import ProductVariantAttributes from "../ProductVariantAttributes";
-import ProductVariantDeleteDialog from "../ProductVariantDeleteDialog";
 import ProductVariantImages from "../ProductVariantImages";
 import ProductVariantImageSelectDialog from "../ProductVariantImageSelectDialog";
 import ProductVariantNavigation from "../ProductVariantNavigation";
@@ -74,135 +73,124 @@ const ProductVariantPage = decorate<ProductVariantPageProps>(
       : undefined;
     return (
       <Toggle>
-        {(isModalActive, { toggle: toggleDeleteModal }) => (
-          <Toggle>
-            {(isImageSelectModalActive, { toggle: toggleImageSelectModal }) => (
-              <>
-                <Container width="md">
-                  <PageHeader title={header} onBack={onBack} />
-                  <Form
-                    initial={{
-                      attributes:
-                        variant && variant.attributes
-                          ? variant.attributes.map(a => ({
-                              slug: a.attribute.slug,
-                              value: a.value.slug
-                            }))
-                          : [],
-                      costPrice:
-                        variant && variant.costPrice
-                          ? variant.costPrice.amount.toString()
-                          : null,
-                      priceOverride:
-                        variant && variant.priceOverride
-                          ? variant.priceOverride.amount.toString()
-                          : null,
-                      quantity:
-                        variant && variant.quantity ? variant.quantity : "",
-                      sku: variant && variant.sku
-                    }}
-                    errors={formErrors}
-                    onSubmit={onSubmit}
-                    key={variant ? JSON.stringify(variant) : "novariant"}
-                  >
-                    {({ change, data, errors, hasChanged, submit }) => (
-                      <>
-                        <div className={classes.root}>
-                          <div>
-                            <ProductVariantNavigation
-                              current={variant ? variant.id : undefined}
-                              variants={
-                                variant
-                                  ? variant.product.variants.edges.map(
-                                      edge => edge.node
-                                    )
-                                  : undefined
-                              }
-                              onRowClick={(variantId: string) => {
-                                if (variant) {
-                                  return onVariantClick(variantId);
-                                }
-                              }}
-                            />
-                          </div>
-                          <div>
-                            <ProductVariantAttributes
-                              attributes={
-                                variant && variant.attributes
-                                  ? variant.attributes.map(a => a.attribute)
-                                  : undefined
-                              }
-                              data={data}
-                              disabled={loading}
-                              onChange={change}
-                            />
-                            <ProductVariantImages
-                              disabled={loading}
-                              images={images}
-                              placeholderImage={placeholderImage}
-                              onImageAdd={toggleImageSelectModal}
-                            />
-                            <ProductVariantPrice
-                              errors={errors}
-                              priceOverride={data.priceOverride}
-                              currencySymbol={
-                                variant && variant.priceOverride
-                                  ? variant.priceOverride.currency
-                                  : variant && variant.costPrice
-                                  ? variant.costPrice.currency
-                                  : ""
-                              }
-                              costPrice={data.costPrice}
-                              loading={loading}
-                              onChange={change}
-                            />
-                            <ProductVariantStock
-                              errors={errors}
-                              sku={data.sku}
-                              quantity={data.quantity}
-                              stockAllocated={
-                                variant ? variant.quantityAllocated : undefined
-                              }
-                              loading={loading}
-                              onChange={change}
-                            />
-                          </div>
-                        </div>
-                        <SaveButtonBar
-                          disabled={loading || !onSubmit || !hasChanged}
-                          state={saveButtonBarState}
-                          onCancel={onBack}
-                          onDelete={toggleDeleteModal}
-                          onSave={submit}
-                        />
-                      </>
-                    )}
-                  </Form>
-                </Container>
-                {variant && (
+        {(isImageSelectModalActive, { toggle: toggleImageSelectModal }) => (
+          <>
+            <Container width="md">
+              <PageHeader title={header} onBack={onBack} />
+              <Form
+                initial={{
+                  attributes:
+                    variant && variant.attributes
+                      ? variant.attributes.map(a => ({
+                          slug: a.attribute.slug,
+                          value: a.value.slug
+                        }))
+                      : [],
+                  costPrice:
+                    variant && variant.costPrice
+                      ? variant.costPrice.amount.toString()
+                      : null,
+                  priceOverride:
+                    variant && variant.priceOverride
+                      ? variant.priceOverride.amount.toString()
+                      : null,
+                  quantity: variant && variant.quantity ? variant.quantity : "",
+                  sku: variant && variant.sku
+                }}
+                errors={formErrors}
+                onSubmit={onSubmit}
+                key={variant ? JSON.stringify(variant) : "novariant"}
+              >
+                {({ change, data, errors, hasChanged, submit }) => (
                   <>
-                    <ProductVariantDeleteDialog
-                      onClose={toggleDeleteModal}
-                      onConfirm={onDelete}
-                      open={isModalActive}
-                      name={variant.name}
-                    />
-                    <ProductVariantImageSelectDialog
-                      onClose={toggleImageSelectModal}
-                      onImageSelect={onImageSelect}
-                      open={isImageSelectModalActive}
-                      images={productImages}
-                      selectedImages={
-                        variant && variant.images && variant.images.edges
-                          ? variant.images.edges.map(edge => edge.node.id)
-                          : undefined
-                      }
+                    <div className={classes.root}>
+                      <div>
+                        <ProductVariantNavigation
+                          current={variant ? variant.id : undefined}
+                          variants={
+                            variant
+                              ? variant.product.variants.edges.map(
+                                  edge => edge.node
+                                )
+                              : undefined
+                          }
+                          onRowClick={(variantId: string) => {
+                            if (variant) {
+                              return onVariantClick(variantId);
+                            }
+                          }}
+                        />
+                      </div>
+                      <div>
+                        <ProductVariantAttributes
+                          attributes={
+                            variant && variant.attributes
+                              ? variant.attributes.map(a => a.attribute)
+                              : undefined
+                          }
+                          data={data}
+                          disabled={loading}
+                          onChange={change}
+                        />
+                        <ProductVariantImages
+                          disabled={loading}
+                          images={images}
+                          placeholderImage={placeholderImage}
+                          onImageAdd={toggleImageSelectModal}
+                        />
+                        <ProductVariantPrice
+                          errors={errors}
+                          priceOverride={data.priceOverride}
+                          currencySymbol={
+                            variant && variant.priceOverride
+                              ? variant.priceOverride.currency
+                              : variant && variant.costPrice
+                              ? variant.costPrice.currency
+                              : ""
+                          }
+                          costPrice={data.costPrice}
+                          loading={loading}
+                          onChange={change}
+                        />
+                        <ProductVariantStock
+                          errors={errors}
+                          sku={data.sku}
+                          quantity={data.quantity}
+                          stockAllocated={
+                            variant ? variant.quantityAllocated : undefined
+                          }
+                          loading={loading}
+                          onChange={change}
+                        />
+                      </div>
+                    </div>
+                    <SaveButtonBar
+                      disabled={loading || !onSubmit || !hasChanged}
+                      state={saveButtonBarState}
+                      onCancel={onBack}
+                      onDelete={onDelete}
+                      onSave={submit}
                     />
                   </>
                 )}
+              </Form>
+            </Container>
+            {variant && (
+              <>
+                <ProductVariantImageSelectDialog
+                  onClose={toggleImageSelectModal}
+                  onImageSelect={onImageSelect}
+                  open={isImageSelectModalActive}
+                  images={productImages}
+                  selectedImages={
+                    variant && variant.images && variant.images.edges
+                      ? variant.images.edges.map(edge => edge.node.id)
+                      : undefined
+                  }
+                />
               </>
             )}
-          </Toggle>
+          </>
         )}
       </Toggle>
     );
