+++ saleor/static/dashboard-next/products/index.tsx
@@ -77,22 +77,20 @@ const Component = () => (
     <Switch>
       <Route exact path={productListUrl()} component={ProductList} />
       <Route exact path={productAddUrl} component={ProductCreate} />
-      <Route exact path={productUrl(":id")} component={ProductUpdate} />
       <Route
         exact
         path={productVariantAddUrl(":id")}
         component={ProductVariantCreate}
       />
       <Route
-        exact
-        path={productVariantEditUrl(":productId", "variantId")}
+        path={productVariantEditUrl(":productId", ":variantId")}
         component={ProductVariant}
       />
       <Route
-        exact
         path={productImageUrl(":productId", ":imageId")}
         component={ProductImage}
       />
+      <Route path={productUrl(":id")} component={ProductUpdate} />
     </Switch>
   </>
 );
