+++ saleor/static/dashboard-next/products/views/ProductImage/urls.ts
@@ -1,0 +1,6 @@
+import * as urlJoin from "url-join";
+
+import { productImageUrl } from "../../urls";
+
+export const productImageRemoveUrl = (productId: string, imageId: string) =>
+  urlJoin(productImageUrl(productId, imageId), "remove");
