+++ saleor/static/dashboard-next/products/views/ProductUpdate/urls.ts
@@ -1,0 +1,5 @@
+import * as urlJoin from "url-join";
+import { productUrl } from "../../urls";
+
+export const productRemoveUrl = (id: string) =>
+  urlJoin(productUrl(id), "remove");
