+++ saleor/static/dashboard-next/products/views/ProductVariant/urls.ts
@@ -1,0 +1,5 @@
+import * as urlJoin from "url-join";
+import { productVariantEditUrl } from "../../urls";
+
+export const productVariantRemoveUrl = (productId: string, variantId: string) =>
+  urlJoin(productVariantEditUrl(productId, variantId), "remove");
